Fixed fist hitting multiple times, added some code for boss to do damage

+++ Assets/Scripts/BossHealth.cs
@@ -7,8 +7,8 @@ public class BossHealth : MonoBehaviour
 {
 
 
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
     public Slider healthSlider; 
 
     void Start()
@@ -23,7 +23,7 @@ public class BossHealth : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         currentHealth -= dmg;
